Change the project's package path to improve its structure
