+++ src/main/java/domain/Bank.java
@@ -1,0 +1,24 @@
+package domain;
+
+public class Bank {
+    private int totalCoin = 282;
+
+    public Bank() {
+    }
+
+    public Bank(int coin) {
+        this.totalCoin = coin;
+    }
+
+    public int getTotalCoin() {
+        return totalCoin;
+    }
+
+    public void payCoin(int coin) {
+        this.totalCoin -= coin;
+    }
+
+    public void gainCoin(int coin) {
+        this.totalCoin += coin;
+    }
+}
