+++ src/main/java/domain/Marketplace.java
@@ -1,0 +1,4 @@
+package domain;
+
+public class Marketplace {
+}
