+++ src/main/java/domain/card/CardType.java
@@ -1,0 +1,11 @@
+package domain.card;
+
+public enum CardType {
+    CROP,
+    Nature_Resources,
+    ANIMAL_HUSBANDRY,
+    SHOP,
+    RESTAURANT,
+    FACTORY_OR_MARKET,
+    MAJOR_ESTABLISHMENT,
+}
