+++ src/main/java/domain/card/establishment/Industry.java
@@ -1,0 +1,8 @@
+package domain.card.establishment;
+
+public enum Industry {
+    BLUE,
+    GREEN,
+    RED,
+    PURPLE
+}
