+++ src/main/java/domain/card/landmark/AmusementPark.java
@@ -1,0 +1,9 @@
+package domain.card.landmark;
+
+import domain.card.CardType;
+
+public class AmusementPark extends Landmark {
+    public AmusementPark() {
+        super("主題樂園", 16, CardType.MAJOR_ESTABLISHMENT, 4);
+    }
+}
