--- src/main/java/main/model/AmusementPark.java
@@ -1,7 +1,0 @@
-package main.model;
-
-public class AmusementPark extends Landmark {
-    public AmusementPark() {
-        super("主題樂園", 16, CardType.MAJOR_ESTABLISHMENT, 4);
-    }
-}
