--- src/main/java/main/model/CardType.java
@@ -1,11 +1,0 @@
-package main.model;
-
-public enum CardType {
-    CROP,
-    Nature_Resources,
-    ANIMAL_HUSBANDRY,
-    SHOP,
-    RESTAURANT,
-    FACTORY_OR_MARKET,
-    MAJOR_ESTABLISHMENT,
-}
