--- src/main/java/main/model/Dice.java
@@ -1,17 +1,0 @@
-package main.model;
-
-import java.util.Random;
-
-public class Dice {
-
-    private int point;
-    private final Random random = new Random();
-
-    public int getPoint() {
-        return point;
-    }
-
-    public void throwDice() {
-        this.point = random.nextInt(6) + 1;
-    }
-}
