--- src/main/java/main/model/Industry.java
@@ -1,8 +1,0 @@
-package main.model;
-
-public enum Industry {
-    BLUE,
-    GREEN,
-    RED,
-    PURPLE
-}
