--- src/main/java/main/model/Marketplace.java
@@ -1,4 +1,0 @@
-package main.model;
-
-public class Marketplace {
-}
