+++ src/test/java/utils/TestUtils.java
@@ -1,0 +1,26 @@
+package utils;
+
+import domain.Bank;
+import domain.Game;
+import domain.Player;
+
+import java.util.Arrays;
+
+public class TestUtils {
+    public static Game createTestGameWithFixedDiceValue(int dicePoint, int numberOfPlayer) {
+        return new Game(new Bank(100), Arrays.asList(createPlayerArrayWithFixedNumber(numberOfPlayer)), null, null) {
+            @Override
+            public int getCurrentDicePoint() {
+                return dicePoint;
+            }
+        };
+    }
+
+    private static Player[] createPlayerArrayWithFixedNumber(int numberOfPlayer) {
+        var players = new Player[numberOfPlayer];
+        for (int i = 0; i < numberOfPlayer; i++)
+            players[i] = new Player(String.valueOf(i + 1));
+
+        return players;
+    }
+}
